refactor: Product 도메인 패키징
