+++ src/main/java/cart/domain/product/ProductPrice.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.product;
 
 import java.util.Objects;
 
@@ -28,8 +28,12 @@ public class ProductPrice {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
         ProductPrice that = (ProductPrice) o;
         return price == that.price;
     }
