+++ src/main/java/cart/dto/ProductsResponse.java
@@ -1,6 +1,6 @@
 package cart.dto;
 
-import cart.domain.Product;
+import cart.domain.product.Product;
 import java.util.List;
 import java.util.stream.Collectors;
