+++ src/test/java/cart/controller/AdminControllerTest.java
@@ -7,7 +7,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.view;
 
-import cart.domain.Product;
+import cart.domain.product.Product;
 import cart.dto.ProductsResponse;
 import cart.service.ProductService;
 import java.util.List;
