+++ src/test/java/cart/domain/ProductNameTest.java
@@ -1,5 +1,6 @@
 package cart.domain;
 
+import cart.domain.product.ProductName;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
